:recycle: - refactor: #104 - separate hook
